Updated abstract - and posed some key questions
Hi Fergus - I have started to revise the abstract. I think this is a great piece of work - but as it stands it will likely be desk rejected from Lancet GH. You need to really convince them in the abstract (and a really well crafted cover letter) exactly why your ms is so critical.
It's a real art.... (and I have by no means mastered it).
I really want to have a good go at editing this but  not sure if this way will work - tell me if it does (or does not!). I don't know why I cannot seem to clone your repo?
I'm sending a message seperately on slack with more info
K